Move move game stuff to external file I know it's weird
